requirements.txt
django-5.0.7
djangorestframework-3.15.2
django-cors-headers
djongo-1.3.6
pymongo-3.12.3
dnspython-2.6.1